// src/GupshupPartnerApiClient/types/Template.ts
export interface Template{
  appId: string;
  category: string;
  createdOn: number;
  data: string;
  elementName: string;
  id: string;
  languageCode: string;
  languagePolicy: string;
  master: boolean;
  meta: string;
  modifiedOn: number;
  namespace: string;
  status: string;
  templateType: string;
  vertical: string;
  reason: string;
}